projects:ad9739a-fmc-ebz: Use iio_app
Refactor project to use common iio_app
