+++ projects/ad9739a-fmc-ebz/src/parameters.h
@@ -48,6 +48,7 @@
 #define TX_DMA_BASEADDR				XPAR_AXI_AD9739A_DMA_BASEADDR
 #define UART_DEVICE_ID				XPAR_XUARTPS_0_DEVICE_ID
 #define UART_IRQ_ID				XPAR_XUARTPS_1_INTR
+#define UART_BAUDRATE                           115200
 #define INTC_DEVICE_ID				XPAR_SCUGIC_SINGLE_DEVICE_ID
 
 #endif /* _PARAMETERS_H_ */
